Player Functionality
Added the basis for drawing the player object, as well as the player input logic to control the player character. Which consists of pressing the space bar to jump, and simulating gravity. Also added the basis for a score system, which as of now is based on deltaTime. Created a class to handle player logic.